chore: create a new component for nested logs
